+++ app.js
@@ -1,5 +1,6 @@
 import { productsBelowTen } from "./apis/belowten.js"
 import { productsOverTen } from "./apis/overten.js"
+import { createTemplate } from "./createTemplates/create-templates.js"
 
 const productsBelow = productsBelowTen['items']
 const productsOver = productsOverTen['items']
@@ -27,30 +28,7 @@ window.onload = () => allItems.map(item =>
 
 /*FUNÇÃO RESPONSAVEL POR CRIAR O TEMPLATE DOS PRODUTOS */
 
-function createTemplate(item) {
 
-    const { name, imageUrl, price, uniqueId } = item
-    const priceTreat = treatPrices(price)
-
-    return `
-       <div class="truffles" id="${uniqueId}">
-            <div class="imagetruffles">
-                <img src="${imageUrl}" alt="">
-            </div>
-            <div class="information">
-                <div>
-                    <h1 class="name-products">${name}</h1>
-                </div>
-                <div>
-                    <span class="price">${priceTreat}</span>
-                </div>
-                <div class="buttons-add-to-cart">
-                    <button class="buttons-action add-to-cart finalize-purchase">Adicionar ao carrinho</button>
-                </div>
-            </div>
-       </div>
-    `
-}
 
 /*FUNÇÃO RESPONSAVEL POR CRIAR O TEMPLATE DOS PRODUTOS DENTRO DO CARRINHO */
